requirements.txt
pyyaml
chromadb
flask
anthropic
chainlit
